+++ amqpstorm/tests/unit/base_tests.py
@@ -79,3 +79,18 @@ class StatefulTests(unittest.TestCase):
         stateful.set_state(Stateful.CLOSING)
 
         self.assertTrue(stateful.is_closing)
+
+    def test_stateful_get_current_state(self):
+        stateful = Stateful()
+
+        stateful.set_state(Stateful.CLOSED)
+        self.assertEqual(stateful.current_state, Stateful.CLOSED)
+
+        stateful.set_state(Stateful.CLOSING)
+        self.assertEqual(stateful.current_state, Stateful.CLOSING)
+
+        stateful.set_state(Stateful.OPEN)
+        self.assertEqual(stateful.current_state, Stateful.OPEN)
+
+        stateful.set_state(Stateful.OPENING)
+        self.assertEqual(stateful.current_state, Stateful.OPENING)